XOR finally works after fixing initialization bugs

+++ Project2-Character-Recognition/character_recognition/mlp.h
@@ -9,37 +9,37 @@ namespace CharacterRecognition {
     Common::PerformanceTimer& timer();
 	class NeuralNet {
 		//The weight matrices
-		vector<float*>w;
+		vector<double*>w;
 		//The bias vectors
-		vector<float*>b;
+		vector<double*>b;
 		// z^[l] = W^[l]a^[l-1] + b^[l]
-		vector<float*>z;
+		vector<double*>z;
 		//a^[l] = g^[l](z^[l])
 		// a^[0] is the input
-		vector<float*>a;
+		vector<double*>a;
 		// The derivatives of weight matrices
-		vector<float*>dw;
+		vector<double*>dw;
 		// The derivatives of bias matrices
-		vector<float*>db;
+		vector<double*>db;
 		// The derivatives of z
-		vector<float*>dz;
+		vector<double*>dz;
 		// The derivatives of a
-		vector<float*>da;
+		vector<double*>da;
 		// The layer sizes
 		vector<int>layer_sizes;
 		// The activation differentiation vectors
-		vector<float*>ghat;
+		vector<double*>ghat;
 	public:
 		//The constructor takes in the input size and the number of output classes
 		NeuralNet(int input_size, int classes, vector<int>layers);
 		// The forward function will do the forward propogation
-		float* forward(float *input);
+		double* forward(double *input);
 		// The backward function is responsible for calculating gradients
 		// and applying the update formula
 		// w^[l] = w^[l] - alpha*dw^[l]
-		void backward(float* y);
+		void backward(double* y);
 		// Calculates the loss given predicted value of y and actual value of y and stores it in loss;
-		//void calculateLoss(float *ypred, float* y, float* loss);
+		//void calculateLoss(double *ypred, double* y, double* loss);
 		// The descrutor will free up the memory;
 		~NeuralNet();
 	};
